feat(crypto): 시장별 수량 정규화 정책 모듈 추가
코인은 8자리 소수점 floor, 주식은 정수 수량 semantics를 유지하는
quantity_policy.py를 도입하고, TradeSignal.suggested_quantity 타입을
int → float로 확장하여 코인 fractional 수량을 end-to-end 지원한다.

+++ strategy/signal.py
@@ -12,7 +12,7 @@ class TradeSignal:
     action: SignalAction
     strength: float  # 0~1
     suggested_price: float | None = None
-    suggested_quantity: int | None = None
+    suggested_quantity: float | None = None
     target_price: float | None = None
     stop_loss_price: float | None = None
     urgency: SignalUrgency = SignalUrgency.WAIT
